+++ benches/medium_table_spec.lua
@@ -8,10 +8,10 @@ describe('medium-table.md', function()
         util.less_than(util.setup('temp/medium-table.md'), 100)
         util.num_marks(base_marks)
 
-        util.less_than(util.move_down(1), 0.5)
+        util.less_than(util.move_down(1), 1)
         util.num_marks(base_marks + 1)
 
-        util.less_than(util.modify(), 7.5)
+        util.less_than(util.modify(), 10)
         util.num_marks(base_marks + 1)
     end)
 end)