+++ benches/readme_spec.lua
@@ -5,13 +5,13 @@ local util = require('benches.util')
 describe('README.md', function()
     it('default', function()
         local base_marks = 114
-        util.less_than(util.setup('README.md'), 40)
+        util.less_than(util.setup('README.md'), 50)
         util.num_marks(base_marks)
 
-        util.less_than(util.move_down(1), 0.5)
+        util.less_than(util.move_down(1), 1)
         util.num_marks(base_marks + 2)
 
-        util.less_than(util.modify(), 5)
+        util.less_than(util.modify(), 10)
         util.num_marks(base_marks + 2)
     end)
 end)