+++ lua/render-markdown/health.lua
@@ -6,7 +6,7 @@ local state = require('render-markdown.state')
 local M = {}
 
 ---@private
-M.version = '8.10.10'
+M.version = '8.10.11'
 
 function M.check()
     M.start('versions')